+++ backend/src/main/java/com/jope/financetracker/service/TransactionService.java
@@ -7,13 +7,16 @@ import com.jope.financetracker.exceptions.DatabaseException;
 import com.jope.financetracker.exceptions.ResourceNotFoundException;
 import com.jope.financetracker.model.Category;
 import com.jope.financetracker.model.Costumer;
+import com.jope.financetracker.model.RecurringTransaction;
 import com.jope.financetracker.model.Transaction;
 import com.jope.financetracker.repository.TransactionRepository;
 
 import org.springframework.dao.DataAccessException;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -24,17 +27,19 @@ public class TransactionService {
     private final TransactionRepository transactionRepository;
     private final CostumerService costumerService;
     private final CategoryService categoryService;
+    private final CurrentUserService currentUserService;
 
     public TransactionService(TransactionRepository transactionRepository, CostumerService costumerService,
-            CategoryService categoryService) {
+            CategoryService categoryService, CurrentUserService currentUserService) {
         this.transactionRepository = transactionRepository;
         this.costumerService = costumerService;
         this.categoryService = categoryService;
+        this.currentUserService = currentUserService;
     }
 
     public Transaction createTransaction(TransactionRequestDTO obj) {
-        Costumer cos = costumerService.findById(obj.costumerId());
-        Category cat = categoryService.findCategoryById(obj.categoryId(), obj.costumerId());
+        Costumer cos = costumerService.findById(currentUserService.getCurrentUserId());
+        Category cat = categoryService.findCategoryById(obj.categoryId(), currentUserService.getCurrentUserId());
         Transaction t = new Transaction();
         t.setCostumer(cos);
         t.setCategory(cat);
@@ -46,10 +51,9 @@ public class TransactionService {
 
     public List<Transaction> createInstallmentsTransaction(InstallmentsTransactionRequestDTO obj) {
         UUID groupId = Generators.timeBasedEpochGenerator().generate();
-        Costumer cos = costumerService.findById(obj.costumerId());
-        Category cat = categoryService.findCategoryById(obj.categoryId(), obj.costumerId());
-        BigDecimal installmentAmount = obj.amount()
-                .divide(BigDecimal.valueOf(obj.installmentTotal()));
+        Costumer cos = costumerService.findById(currentUserService.getCurrentUserId());
+        Category cat = categoryService.findCategoryById(obj.categoryId(), currentUserService.getCurrentUserId());
+        BigDecimal installmentAmount = obj.amount().divide(BigDecimal.valueOf(obj.installmentTotal()), 2, RoundingMode.HALF_UP);
 
         List<Transaction> transactionsToSave = new ArrayList<>();
 
@@ -61,8 +65,8 @@ public class TransactionService {
             transaction.setCostumer(cos);
             transaction.setCategory(cat);
             transaction.setAmount(installmentAmount);
-            transaction.setDescription(obj.description() + " (Parcela " + i + " de " + obj.installmentTotal() + ")");
-            transaction.setDate(obj.date().plusMonths(i - Long.valueOf(1)));
+            transaction.setDescription(obj.description() + " (Installment " + i + " of " + obj.installmentTotal() + ")");
+            transaction.setDate(obj.date().plusMonths(i - 1L));
             transactionsToSave.add(transaction);
         }
 
@@ -73,7 +77,8 @@ public class TransactionService {
         if (groupUUID == null) {
             throw new IllegalArgumentException("Installments transaction group id must not be null!");
         }
-
+        List<Transaction> list = transactionRepository.findAllByInstallmentGroupId(groupUUID);
+        list.forEach(x -> currentUserService.checkAccess(x.getCostumer().getId()));
         try {
             transactionRepository.deleteByInstallmentGroupId(groupUUID);       
         } catch (DataAccessException ex) {
@@ -82,18 +87,19 @@ public class TransactionService {
     }
 
     public Transaction getTransactionById(Long id) {
-        return transactionRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(id));
+        Transaction transaction = transactionRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
+        currentUserService.checkAccess(transaction.getCostumer().getId());
+        return transaction;
     }
 
     public List<Transaction> getAllTransactions() {
-        return transactionRepository.findAll();
+        return transactionRepository.findAllByCostumerId(currentUserService.getCurrentUserId());
     }
 
     public Transaction updateTransaction(Long id, Transaction transactionDetails) {
         Transaction transaction = getTransactionById(id);
+        currentUserService.checkAccess(transaction.getCostumer().getId());
 
-        transaction.setCostumer(transactionDetails.getCostumer());
         transaction.setCategory(transactionDetails.getCategory());
         transaction.setAmount(transactionDetails.getAmount());
         transaction.setDate(transactionDetails.getDate());
@@ -103,6 +109,12 @@ public class TransactionService {
     }
 
     public void deleteTransaction(Long id) {
-        transactionRepository.deleteById(id);
+        Transaction b = transactionRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Transaction not found!"));
+        currentUserService.checkAccess(b.getCostumer().getId());
+        try {
+            transactionRepository.deleteById(id);
+        } catch (DataAccessException ex) {
+            throw new DatabaseException("Failed to delete transaction: " + id);
+        }
     }
 }
